relocate file to the correct location
